fix: naming, option chain and folder structure
